# Conflicts:
#	backend/node_modules/.package-lock.json
#	backend/package-lock.json

+++ backend/controllers/eventController.js
@@ -1,6 +1,8 @@
 const eventSchema = require('../models/eventModel');
 const managerSchema = require('../models/managerModel');
 
+const {getUserCurrentTime} = require('../middleware/userTime');
+
 const mongoose = require('mongoose');
 
 
@@ -9,8 +11,22 @@ const createEvent = async(req, res) => {
     try {
         let user_id = res.locals.user._id.toString();
         req.body.host_id = user_id;
+        
+        const now = getUserCurrentTime();
+        console.log(now);
 
+        
+        if (req.body.startTime < now) {
+            return res.status(400).json({mssg: 'Cannot an event in the past'});
+        }
+
+        if (req.body.startTime > req.body.endTime) {
+            return res.status(400).json({mssg: 'End time must be later than start time'});
+        }
+
+        req.body.status = 'upcoming';
         const event = await eventSchema.create(req.body);
+        
         const manager = await managerSchema.findOne({user_id: res.locals.user._id});
         
         manager.events.push({
@@ -22,26 +38,37 @@ const createEvent = async(req, res) => {
             member_id: user_id
         })
 
-
-
         manager.save();
         event.save();
 
         res.status(200).json({event, manager});
     } catch (error) {
-        res.status(400).json({error: error.message});
+        res.status(400).json({mssg: 'Cannot create event', error: error.message});
     }
 }
 
 //Get all events
 const getEvents = async(req, res) => {
     try {
-        const event = await eventSchema.find({});
+        const { title, startTime, endTime, status } = req.body; //status: upcoming, occuring, ended
+        let query = {};
+
+        if (title) {
+            query.title = new RegExp(title, 'i');
+        }
+
+        if (startTime && endTime) {
+            query.startTime = { $gte: new Date(startTime) };
+            query.endTime = { $lte: new Date(endTime) };
+        }
+
+        const event = await eventSchema.find(query);
         res.status(200).json({event});
     } catch (error) {
         res.status(404).json({error: error.message});
     }
 }
+
 
 //Get event by ID
 const getEventByID = async(req, res) => {
@@ -61,12 +88,12 @@ const getEventByID = async(req, res) => {
 }
 
 //Update event
-const updateEvent = async(req, res) => {
+const updateEvent = async (req, res) => {
     const {id} = req.params;
 
     try {
-        const event = await eventSchema.findOneAndUpdate({_id: id}, {...req.body});
-
+        const event = await eventSchema.findOneAndUpdate({_id: id}, {...req.body}, {new: true});
+        f
         if (!mongoose.Types.ObjectId.isValid(id)) {
             return res.status(404).json('Cannot find event');
         }
@@ -82,13 +109,30 @@ const deleteEvent = async(req, res) => {
     const {id} = req.params;
 
     try {
-        const event = await eventSchema.findOneAndDelete({_id: id});
+        //Remove event from all user's event manager
+        const event = await eventSchema.findOne({_id: id});
 
-        if (!mongoose.Types.ObjectId.isValid(id)) {
-            return res.status(404).json('Cannot find event');
+        if (event.host_id === res.locals.user._id.toString()) {
+            await Promise.all(event.member_list.map(async (member) => {
+                await managerSchema.findOneAndUpdate(
+                    {user_id: member.member_id},
+                    {$pull: {events: {event_id: id}}},
+                    {new: true}
+                );
+            }));
+            
+            //Remove event
+            await eventSchema.findOneAndDelete({_id: id});
+    
+            if (!mongoose.Types.ObjectId.isValid(id)) {
+                return res.status(404).json('Cannot find event');
+            }
+
+            return res.status(200).json({mssg: 'Deleted event'});
         }
 
-        res.status(200).json({event});
+        res.status(400).json({mssg: 'User have no permission to delete this event'});
+
     } catch (error) {
         res.status(400).json({error: error.message});
     }
@@ -102,6 +146,10 @@ const joinEvent = async(req, res) => {
     try {
         //Add member_id to event's member list
         const event = await eventSchema.findById(event_id);
+
+        if (event.member_list.length > event.capacity && event.capacity > 0) {
+            return res.status(400).json({mssg: 'Event is full'});
+        }
 
         //Check if user already in event
         const flag = event.member_list.find(tmp => tmp.member_id === member_id.toString());
@@ -131,6 +179,31 @@ const joinEvent = async(req, res) => {
     }
 }
 
+//Leave event
+const leaveEvent = async (req, res) => {
+    try {
+        const {id} = req.params;
+        const user_id = res.locals.user._id;
+    
+        //Remove user from event
+        await eventSchema.findOneAndUpdate(
+            {_id: id},
+            {$pull: {member_list: {member_id: user_id}}}
+        );
+    
+        //Remove event from user's manager
+        await managerSchema.findOneAndUpdate(
+            {user_id: user_id},
+            {$pull: {events: {event_id: id}}},
+            {new: true}
+        );
+
+        res.status(200).json({mssg: 'Left event'});
+    } catch (error) {
+        res.status(400).json(error.message);
+    }
+}
+
 
 module.exports = {
     createEvent, 
@@ -138,7 +211,8 @@ module.exports = {
     getEventByID,
     getEvents,
     updateEvent,
-    joinEvent
+    joinEvent,
+    leaveEvent
 }
